Refactor core systems, math utilities, and rendering pipeline; improve hashing, transforms, and ECS integration.
Introduced Handle for strongly-typed resource identification and updated allocation API.
Enhanced hash support for heterogeneous types and fixed Pool::Clear() behavior.
Added Bit::BitCast for zero-overhead type conversions.
Reworked Matrix3x2 to scalar representation.
Implemented Matrix3x2 transform operations for Circle, Line, Quad, and Rect.
Integrated origin handling directly into transform matrices.
Added skew transformations to Matrix3x2 and Transform2D.
Fixed namespace inconsistencies across math classes.
Improved sprite and text rendering.
Fixed line rendering issues.
Embedded font effect ID into Matrix2x3Packed (preparing for further compression).
Removed Content::Proxy (no longer needed with ECS archetypes).
Added templated overloads for pair DSL types in Entity.
Updated to latest Flecs version using the new Range API.
Added serialization support for Pool, Slot, and Catalog.

+++ Engine/Public/Zyphryon.Base/Text.hpp
@@ -65,7 +65,7 @@ inline namespace Base
     static void IterateUTF8(ConstStr8 Text, AnyRef<Function> Callback)
     {
         ConstPtr<Char> Data      = Text.data();
-        UInt           Size      = Text.size();
+        size_t         Size      = Text.size();
         UInt32         Codepoint = SDL_StepUTF8(&Data, &Size);
 
         while (Codepoint != 0)
